service + add website + inputs + session

+++ src/lib/sessionStorage.ts
@@ -1,5 +1,7 @@
 import { Key } from "react";
 
+const isBrowser = typeof window !== 'undefined';
+
 export enum SKey {
     Projects = 'projects',
     Categories = 'categories',
@@ -8,16 +10,19 @@ export enum SKey {
 
 export default class SessionStorage {
     public static get = (key: Key) => {
+        if (!isBrowser) return undefined;
         const value = sessionStorage.getItem(key as string);
         return value ? JSON.parse(value) : undefined;
     };
 
     public static set = (key: Key, anyValue: string | number | Object) => {
+        if (!isBrowser) return;
         const value = JSON.stringify(anyValue);
         sessionStorage.setItem(key as string, value);
     };
 
     public static setAll = (valuesArray: [Key, string | number | Object][]) => {
+        if (!isBrowser) return;
         valuesArray.forEach((value) => {
             this.set(value[0], value[1]);
         });
@@ -43,10 +48,12 @@ export default class SessionStorage {
     };
 
     public static remove = (key: Key) => {
+        if (!isBrowser) return;
         sessionStorage.removeItem(key as string);
     };
 
     public static clear = () => {
+        if (!isBrowser) return;
         sessionStorage.clear();
     };
 }
